fix(ci): install project package in github actions
The github actions workflow was failing due to a `ModuleNotFoundError`.
This was because the `snyker` package itself was not being installed,
only its dependencies.

This change removes the `--no-root` flag from the `poetry install`
command so that the `snyker` package is installed and the tests can
find and import it.